move the deploy contract method to common package
